services/ProjectManagementService/requirements.txt
djangorestframework
markdown
django-filter
pytest-django